BaseLine_BE_main/requirements.txt
fastapi
uvicorn
psycopg2-binary
python-dotenv
pydantic
google-cloud-aiplatform
google-adk
supabase
websockets
google-auth
requests